EJ 4
terminado

+++ 8-TPs/8-TPs/jsFerreteConstruccion.js
@@ -3,6 +3,7 @@ A.	 Mostrar la cantidad de alambre a comprar  si se ingresara el largo y el anch
 B.	Mostrar la cantidad de alambre a comprar  si se ingresara el radio  de un terreno circular y se debe alambra con tres hilos de alambre.
 C.	Para hacer un contrapiso de 1m x 1m se necesitan 2 bolsas de cemento y 3 de cal, debemos mostrar cuantas bolsas se necesitan de cada uno para las medidas que nos ingresen.
 */
+
 var largo;
 var ancho;
 var perimetro;
@@ -29,8 +30,8 @@ function Circulo ()
 }
 function Materiales () 
 {
-	largo=parseInt(document.getElementById("Largo").value);
-    ancho=parseInt(document.getElementById("Ancho").value);
+	largo=parseFloat(document.getElementById("Largo").value);
+    ancho=parseFloat(document.getElementById("Ancho").value);
     superficie= largo*ancho;
     cemento= superficie*2;
     cal=superficie*3;
